requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0